+++ src/main/java/br/com/luizsn/VassCommerce/model/Cliente.java
@@ -8,7 +8,7 @@ public class Cliente extends  Usuario{
     private LocalDate dataNascimento;
     private String cpf;
 
-    public Cliente(int id,String nomeCompleto,String email,String senha,Date dataCadastro,Date dataUltimaAtualizacao,String fotoUrl,
+    public Cliente(long id,String nomeCompleto,String email,String senha,Date dataCadastro,Date dataUltimaAtualizacao,String fotoUrl,
                    LocalDate dataNascimento,String cpf) {
         super(id,nomeCompleto,email,senha,dataCadastro,dataUltimaAtualizacao,fotoUrl);
         this.dataNascimento =dataNascimento;
